fix(openapi): widen counter to int64 + reject downgraded SK flags
Three more Codex P2 review items on #18:

- README install snippets pinned to v0.1.0 (Go and Swift); bump to
  v0.2.0 so readers following the docs after the tag land on a schema
  that actually carries `counter`/`flags`/`ssh_sk_flags`. Also future-
  proof the "v0.1.0 series" phrasing in the known-debt section to
  "v0.x series".
- Add `format: int64` to the SSH success-response `counter` field so
  oapi-codegen emits `int64` instead of `int`. Go's `int` is 32-bit on
  GOARCH=386 / arm, which cannot represent the full u32 counter range
  the schema allows (and would have made the new max-boundary Go
  regression test uncompilable on those targets). The description now
  spells out that the on-wire value is u32 and instructs receivers to
  serialise the byte slice as 4 big-endian bytes regardless of the
  language-side representation.
- Tighten the per-signature `flags` description to require approvers
  to either assert with at least the requested bits or return a
  failure response — a success response whose asserted flags clear
  bits the requester set would silently downgrade the security
  posture (e.g. UV-required → UP-only). Receivers SHOULD additionally
  validate the asserted byte against the request as belt-and-
  suspenders defence against a misbehaving approver.

Regenerated bindings. Existing regression tests already cover the new
generated `int64` type since the Go literals (`7`, `42`, `4294967295`,
`255`) are valid `int64` constants. `make generate-check` clean;
`make test` green (Go + Swift + TS).

+++ go/types.gen.go
@@ -758,10 +758,10 @@ type MailboxSshAuthResponseSuccessV1 struct {
 	// ApprovalProof Canonical Longfellow approval proof carried inside encrypted approval responses.
 	ApprovalProof *ApprovalAttestedKeyProof `json:"approval_proof,omitempty"`
 
-	// Counter Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
-	Counter int `json:"counter"`
+	// Counter Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
+	Counter int64 `json:"counter"`
 
-	// Flags Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
+	// Flags Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
 	Flags int `json:"flags"`
 
 	// Signature RFC 4648 standard base64 with `=` padding for the raw SSH signature blob (no SSH-wire framing).
@@ -814,10 +814,10 @@ type MailboxSshSignResponseSuccessV1 struct {
 	// ApprovalProof Canonical Longfellow approval proof carried inside encrypted approval responses.
 	ApprovalProof *ApprovalAttestedKeyProof `json:"approval_proof,omitempty"`
 
-	// Counter Monotonic counter the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)`. Successive signatures from the same key handle MUST have strictly increasing counter values.
-	Counter int `json:"counter"`
+	// Counter Monotonic counter (u32) the signer's secure element returned for this SK signing operation. Receivers MUST embed this in the OpenSSH SK signature preimage at the position between `flags` and `SHA256(data)` as a 4-byte big-endian unsigned integer. Successive signatures from the same key handle MUST have strictly increasing counter values. The schema declares `format: int64` so 32-bit Go targets can still represent the full u32 range without overflow.
+	Counter int64 `json:"counter"`
 
-	// Flags Per-signature SK assertion flags byte the signer's secure element actually asserted with. MAY differ from the request `flags` byte (e.g. when the SK could not deliver user verification it MUST clear `0x04` on the assertion). Receivers MUST embed this byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead.
+	// Flags Per-signature SK assertion flags byte the signer's secure element actually asserted with. Approvers MUST either (a) assert with at least the bits the request `flags` byte asked for (UP=0x01, UV=0x04) and return the resulting byte here, or (b) return a `MailboxSshAuthResponseFailureV1` / `MailboxSshSignResponseFailureV1` with the appropriate signing error code. Approvers MUST NOT return a success response whose asserted flags byte clears bits the requester set; that would silently downgrade the security posture (e.g. UV-required → UP-only) below what the request agreed to. Receivers MUST embed this asserted byte at the `flags` position of the OpenSSH SK signature preimage; verification fails if the request `flags` byte is used instead. Receivers SHOULD additionally verify that every bit set in the request `flags` byte is also set here as belt-and-suspenders defence against a misbehaving approver.
 	Flags int `json:"flags"`
 
 	// Signature RFC 4648 standard base64 with `=` padding for the raw SSH signature blob (no SSH-wire framing).
